requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
prophet==1.1.5
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
